feat(request): add `parseBody()` method. body-parse middleware is obsolete
Close #362
