+++ README.md
@@ -290,7 +290,7 @@ app.use('*', logger())
 app.use('/posts/*', cors())
 
 // specify method and path
-app.post('/posts/*', basicAuth(), bodyParse())
+app.post('/posts/*', basicAuth())
 ```
 
 If the handler returns `Response`, it will be used for the end-user, and stopping the processing.
@@ -302,7 +302,7 @@ app.post('/posts', (c) => c.text('Created!', 201))
 In this case, four middleware are processed before dispatching like this:
 
 ```ts
-logger() -> cors() -> basicAuth() -> bodyParse() -> *handler*
+logger() -> cors() -> basicAuth() -> *handler*
 ```
 
 ### Built-in Middleware
@@ -424,6 +424,12 @@ app.get('/entry/:id', (c) => {
 // Parse cookie
 app.get('/entry/:id', (c) => {
   const value = c.req.cookie('name')
+  ...
+})
+
+// Parse Request body
+app.post('', (c) => {
+  const body = c.req.parseBody()
   ...
 })
 ```
